Reprise du projet

+++ core/src/com/guco/tap/entity/Weapon.java
@@ -4,7 +4,7 @@ public class Weapon {
     public String icon;
     public String name;
     public int lvl;
-    public int damage_value;
+    public float damage_value;
     public int damage_currency;
     public int critical_value;
     public int critical_currency;
